refactor: 메뉴 생성 수정
