+++ src/main/java/kitchenpos/domain/Menu.java
@@ -10,7 +10,6 @@ import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
 import javax.persistence.OneToMany;
-import kitchenpos.application.MenuValidator;
 
 @Entity
 public class Menu {
